renaming block files
