+++ js/main.js
@@ -1,5 +1,5 @@
-
 import renderScreen from './render-screen';
+
 
 const questionData1 = {
   type: `double`,
