+++ js/render-screen.js
@@ -1,5 +1,5 @@
 import getIntro from './screen-intro';
-import getGreeting from '../screen-greeting';
+import getGreeting from './screen-greeting';
 import getRuls from './screen-rules';
 import getGame from './screen-game';
 import getStats from './screen-stats';
