+++ js/screen-rules.js
@@ -1,7 +1,7 @@
 import renderTemplate from './render-template';
 import controller from './controller';
-import getFooter from './footer';
-import getHeader from './header';
+import getFooter from './block-footer';
+import getHeader from './block-header';
 
 
 const contentTemplate = `
